Export conda environment variables
These are read by some packages in order to find
the location of compiled libraries etc.

Normally these variables are exported by .bashrc
but since we don't run that in our container entrypoint
the variables are not defined during container runtime
without this.

This fixes `import bitsandbytes` on gpu instances with
cudatoolkit