Adds support for 20G A100s variants (#257)
This adds client support for smaller A100s with 20Gb instead of the default 40gb. I decided to go with the following interface:

```python
@stub.function(gpu=modal.gpu.A100(memory=20))  # or "a100-20g"
def f(): ...
```

That’ll fail if memory is not in (0, 20, 40). My other option is to create a new class (`A100_20g`). I found that confusing because the trailing `G` on `A10G` does not mean "gigabytes". It also feels repetitive to create a class per memory variant. 

`memory` also supports the higher memory variant (80Gb) whenever that is available on Modal.