Refactor private registry protos
The private registry configs should not go in the `BaseImage` but in the
`Image` definition. That's the base because `BaseImage` is an internal
reference and not a mapping to a Docker `FROM` command.

I aslo refactored the interface a bit to make it easier to read.

`BaseImage` registry fields are now deprecated.