Quickfix: Don't mount the origin package for each function more than once (#555)
* Don't mount the origin package for each function more than once

Regardless if there are multiple functions in the stub. This patch simply
reintroduces the _function_mounts stub-level mount cache that we had until
last week.

It's a bit janky, but we should refactor this mount logic soon and get rid
of a bunch of this. Preferably in a way that lets us upload all mount files
in one common thread pool instead of using one for each mount sequentially

* Don't use stub for deduplication if there is none