Add experimental volumes API
This is API for working with volumes, which are durable storage units
exposed via the file system. Unlike a networked file system changes
to the volumes are only made visible when explcitly committed.
Similarly, to get updates to the volume you need to explicitly reload.

Note that the backend support for volumes is not yet in place, so they
are non-functional as is.