Use an empty environment name by default
Instead, we'll treat an unspecified environment (empty string) on the backend by assuming there is only one environment and using that one, or error if there are multiple ones.

This lets the users have any name on their 'single' environments, and legacy clients will use that one. When a workspace has multiple environments, the profile (or cli command) should always explicitly state which one to use, to avoid mishaps.