Make "modal" the official client package
Currently, the official client package is named "modal-client"
with "modal" as a thin wrapper or alias. This change makes it so that
"modal" is the official package on PyPI, and we leave "modal-client" as
the name of the legacy compatibility package.

This also updates the README and other references to use the new name.