Do not allow large serialized functions
Modal breaks with a cryptic "unknown 413" gRPC error message when a function in a Jupyter Notebook closes over a global variable in scope that is over 100 MiB when serialized. This PR warns users if their function closes over 256 KiB of data (a conservative limit), then emits a full error if the size of over 16 MiB.

Also added a test for this behavior in the client.