Adds checkpoint / restore logic to container entrypoint (#1007)
A checkpointing task will execute a function and, after its import sequence, it will call the gRPC method `ContainerCheckpoint` and enter a [busy-wait loop](https://en.wikipedia.org/wiki/Busy_waiting) waiting for the environment variable `MODAL_FUNCTION_RESTORED` to be populated. We execute the target function (or a shell) after the env var is populated.

The call to `ContainerCheckpoint` triggers the checkpointing of the function. The function is checkpointed while busy waiting, so it breaks from that loop when restored.