Remove task-cancellation log spam already cancelled (#1143)
Noticing a lot of log spam of the following sort for ASGI apps, after
upgrading modal-client.

  2023-12-20T19:43:01+0000 Canceling remaining unfinished task <Task
  cancelling name='Task-44'
  coro=<asgi_app_wrapper.<locals>.fn.<locals>.fetch_data_in() running at
  /pkg/modal/_asgi.py:30> wait_for=<Future pending
  cb=[Task.task_wakeup()]> cb=[set.discard()]>

This started happened 0b5f834 when we stopped using grace period in the
surrounding task context. So we stopped running the unfinished (but
marked-for-cancellation) tasks on the event loop when closing the
TaskContext. These tasks were therefore not done(), and we logged the
above. We can check for pending cancellations using the task object - so
just do that before logging.