Updates local state when restoring containers  (#1229)
In order to support ephemeral app checkpointing and restore operations, we need to update FunctionIOManager's local variables `task_id` and `function_id`. Those are used to get inputs for a function.

This is not needed for deployed apps because the function ID doesn't change between checkpointing and restore.

This eventually allows for a checkpoint created with an ephemeral function to be used in restore operations by deployed apps.