[MOD-2209] Add clobber flag for volume uploads
Previously, we silently overwrite any existing files. This is
error-prone so we're changing the default behaviour to not
overwrite existing files.

This adds a `clobber` flag which you can use to explicitly allow
overwriting existing files.

Even with `clobber=True` we will not allow overwriting existing
_directories_ with uploaded files.

The name `clobber` is borrowed from standard unix commands such as
`mv` and `cp`.