Gracefully abort output polling when local app shuts down
Gets rid of the ugly cancellation error muting we did before, in favor of
using an asyncio.Event that is fired when the local app stops either for
a sigint/KeyboardInterrupt *or* due to an app_done signal being received
in the logs.

This fixes the massive tracebacks that were previously spit out when someone
stopped the app in the web interface while having a `modal run` CLI session running.