Do not warn when TaskContext cancels unfinished tasks
See https://github.com/modal-labs/modal-client/pull/1522#issuecomment-1999782348

The history of this is that @irfansharif silenced the warning for Python versions up to 3.10 and below, two months ago. But it's still been happening for version 3.11, and it's a bit of an unhelpful warning in our case.

Note: `asyncio.TaskGroup` from the Python standard library waits for all tasks to finish when it exits normally. It only cancels tasks when it sees that one of its subtasks throws an exception other than `CancelledError`. However, our `async_utils.TaskContext` object is quite different in behavior, since it actually cancels _all_ tasks when it exits after a short grace period.