Working blocking map implementation
Moves the input iterator consumption to the main thread to prevent
it from possibly locking up the synchronicity thread while next() blocks.

This is implemented making `map()` a non-synchronized call and instead
running a temporary event loop in the main thread which builds up an
asyncio.Queue bound to the synchronizer thread with all the input items.