Unify FileEntry protobufs and add a public dataclass (#1680)
* Unify FileEntry protobufs and add a public dataclass

This unifies the SharedVolumeListFilesEntry and VolumeListFilesEntry protobufs, which now just both become FIleEntry. I'm also adding a public dataclass for FileEntry, which is used the API for listdir() on both NFS and Volume objects.

Note that the NFS implementation doesn't have support for symlinks yet, or for the mtime / size fields. I quickly added support for those in server code.

* Python 3.8 syntax

* Fix type annotations

* Fix ruff

* Add deprecation error on __getattr__