Make sure serialized partial functions are actually usable
Synchronicity wrapped objects that are serialized are not generally useful when deserialized in
a different process. If we however add a custom serialization format that pickles the underlying
object and then translates it on the receiving end, we get something that can be used.

This is necessary to get usable methods on classes, since those are PartialFunctions which
are synchronicity wrapped