Fix Cls.lookup() not working after update
Lookups were breaking since backend isn't returning class_function_id as part
of the handle metadata for the class.

Instead of fixing the backend, this removes the class_function_id field
from ClassMetadata and instead does an explicit lookup for the function named <class-tag>.*.
This is in line with an upcoming refactor that could remove the need for synced Cls objects altogheter, since
the "class servicer function" is all that is needed to invoke the class now
anyway.

Client tests were passing since the mock servicer was more feature-complete than the real
backend api.