Adds ProxyGet to public API
Introduces `ProxyGet` to public API. This preceedes an update to
`Proxy.from_name("my-proxy")` to replace `ProxyGetOrCreate`. I indended
to have that RPC support both the client get method and internal proxy
create methods but came to the conclusion that it is better to separate
those methods because it is not desirable to create a Proxy lazily.
(That's the case because it takes several seconds to create atm.)