[CLI-229] Fix for `self.method.remote` calling FunctionBindParams in containers (#2567)
**Note** This only fixes the special case of un-parameterized classes calling stuff on self when they could simply use the class servicer function as is done when using the class locally.

To fix the general case, we'd need to add a way to hydrate parameter-bound instances of classes ("instance service functions") using the container args or similar.

The root cause of the problem was that when used in side containers, the class isn't hydrated at import time when the "self" is created (unlike when it's used in an active app locally). This means that when the Cls is instantiated (in get_user_class_instance) it doesn't trigger the short-circuit logic in _bind_parameters which prevents FunctionBindParams for parameter-less functions. This simple patch moves that short-circuit logic into the _load function instead.