Avoid sending empty messages on process start
Previously, we added an empty message on container start so that the
client could tell the process had booted. This turned out to be a
breaking change since clients would now receive an extra message on
initialization. This commit refactors output consumption to only return
the empty message in `modal shell` usage.