Exclusively start using the FunctionData form ...
.. when creating functions. We were already using the FunctionData form
when creating functions that use GPU fallbacks. On the server, when
trying to create functions using the Function form, we automatically
translate it to the FunctionData form. By starting to do this in the
client, we can start actually moving some of the configuration from the
lower api_pb2.Function proto to the upper api_pb2.FunctionData proto for
good. We'll still have a period where we need to support the
duplication, because of older clients. But the fields can now be treated
as being deprecated.