Fix for nested map calls deadlocking
Bug was caused by obscure bug where the iterable produced by map
was treated as an async generator but it's actually an AsyncOrSyncIterable
custom type. This type lacked the .aclose() method that was expected to
exist which caused an error when its subsequently called.

Will separately investigate why this error was never surfaced as a traceback
which is an indication of something wrong in the async execution stack