Client now uses SYNC for .map() (#2734)
Part of SVC-180.

This PR has the client start using SYNC for .map calls, and gives it the ability to do client retries. We already did this to .remote() in PR #2403.

For now, client retries are disabled by default. They can be enabled per-workspace by adding the workspace id to the CLIENT_RETRY_ENABLED_WORKSPACES in function_fast.py.