Wrap user_cls in Cls inside containers when no Cls available
Uses metadata from the container args to hydrate the class service function
in order to construct a Cls that can then be used to create a more functional
Obj.

Previously, the Cls inside a container for a serialized or locally decorated
class was the actual user class, so it wouldn't have any of the Cls functionality,
notably lacking the synthetic constructor required for parameter-based construction.

This should also fix the fact that self.other_method weren't callable for serialized
classes before