Classes: hydrate service function using container args [CLI-326, CLI-289] (#3008)
* Adds container test that fails due to params spec missing

* Wrap user_cls in Cls inside containers when no Cls available from global decorator (serialized or local decoration)

Uses metadata from the container args to hydrate the class service function
in order to construct a Cls that can then be used to create a more functional
Obj.

Previously, the Cls inside a container for a serialized or locally decorated
class was the actual user class, so it wouldn't have any of the Cls functionality,
notably lacking the synthetic constructor required for parameter-based construction.

This should also fix the fact that self.other_method weren't callable for serialized
classes before

* is another app to avoid lazy

* Hydrate both the class and the underlying class service function

* Fix app layouts for existing tests

* Types

* Update docstring