[prototype] modal torchrun wrapper
this prototypes a wrapper for multinode jobs that allows the
user to specify in a more modal-esque way the function they want
to run. we handle the torchrunning behind the scenes.

this is 100% ai coded so the implementation is a bit jank, but
i really like the interface here! it seems like a natural way
to extend the modal way of doing things to multi-node.