Set force_jupyter=False on rich Console objects
Rich's autodetection of Jupyter is a bit aggressive and changes the way interactive CLI outputs are displayed in notebooks, causing problems while rendering. It appears that just using the terminal output mode is better, and the only way to do this seems to be by "disabling" Jupyter with `force_jupyter=False`.