MVP: `modal launch machine` (#3419)
* MVP: `modal launch workspace`

This is a starting point for a command that just launches an instance with a mounted volume and allows you to SSH into it.

It's meant to be a basic primitive with reasonably good developer experience. The SSH key is generated randomly / ephemerally.

It relies on the notebook base image, so users don't need to build the image from scratch on first usage.

Resolves SPR-142.

* Uppercase GPU

* Fix 10 second wait with hack

* Fix CI, Optional[] in Python 3.9

* Rename `modal launch workspace` to machine

* Add charles' reference