Adds support for cbor2 serialization for function calling [SDK-584] (#3474)
Deployed functions will now support both pickle and cbor inputs, and try to mirror the output format with the input format when allowed.
In addition, functions will now advertise which input and output formats they support to allow a lookup to introspect this.

This enables future versions of modal clients (in particular libmodal clients in other languages) to use cbor instead of pickle for payloads.